* reverseOneDay가 0인 문제로 인하여, 오리들의 수치가 변하지 않는 문제 수정

+++ DuckFarm/Assets/Scripts/World.cs
@@ -9,7 +9,7 @@ public class World : MonoBehaviour
     public static long CurrentGameWorldTimeMS = 0;
 
     public static int oneDay = 2000; //1일 = 2000 Millisecond
-    public static int reverseOneDay = 1000 / oneDay;
+    public static float reverseOneDay = 1000.0f / oneDay;
     protected World()
     {
         _instance = this;
